Rajout de quelques contrôles de saisies dans ajout ville/personne/connexion

+++ include/pages/ajouterVille.inc.php
@@ -17,15 +17,21 @@
   }
   else
   {
-    $ville = new Ville( array('vil_nom' => $_POST["vil_nom"] ));
-    $manager->add($ville);
-    $file = fopen("./log/covoiturage.log","a");
-    fputs($file, "<span>".date('l jS \of F Y h:i:s A')." : Le pc ".$_SERVER["REMOTE_ADDR"]." a affiché la ville".$_POST["vil_nom"]."</span><br>\n");
-    echo "<img src='./image/valid.png' alt='valid'>la ville <b>".$_POST["vil_nom"]."<b> a ete ajoutée ";
-    echo "<br><span id='chrono'>Redirection dans 2 secondes</span>";
-    echo "<script>appel();
-    redirectionAccueil();</script>";
-
+    if((strlen($_POST["vil_nom"]>0))&&(!$manager->existeVille($_POST["vil_nom"]))){
+      $ville = new Ville( array('vil_nom' => $_POST["vil_nom"] ));
+      $manager->add($ville);
+      $file = fopen("./log/covoiturage.log","a");
+      fputs($file, "<span>".date('l jS \of F Y h:i:s A')." : Le pc ".$_SERVER["REMOTE_ADDR"]." a affiché la ville".$_POST["vil_nom"]."</span><br>\n");
+      echo "<img src='./image/valid.png' alt='valid'>la ville <b>".$_POST["vil_nom"]."<b> a ete ajoutée ";
+      echo "<br><span id='chrono'>Redirection dans 2 secondes</span>";
+      echo "<script>appel();
+      redirectionAccueil();</script>";
+    }else{
+      echo "<img src='./image/erreur.png' alt='valid'> Nom de ville nul ou ville déjà existante";
+      echo "<br><span id='chrono'>Redirection dans 2 secondes</span>";
+      echo "<script>appel();
+      redirectionAccueil();</script>";
+    }
   }
